requirements.txt
google-generativeai
openai
requests
alpaca-py